pyproject.toml
[tool.poetry]
name = "hero-test-api"
version = "0.1.0"
description = ""
authors = ["Ira <ira.ivashko.99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
httpx = "^0.28.0"
uvicorn = "^0.32.1"
flake8 = "^7.1.1"
black = "^24.10.0"
pytest-asyncio = "^0.24.0"
pytest-mock = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
